feat(team): move configure wizard into agentic-team configure subcommand (Unit 1)
Move cross-harness team-config wizard out of Pi/omp-branded bin/agentic-configure
into the harness-neutral bin/agentic-team as the configure subcommand.

- Add configure subcommand to agentic-team with identical CLI surface:
  agentic-team configure [--web] [--non-interactive] [--assign role=harness:model ...]
  [--default-harness H] [--path PATH]; exit codes 0/2 preserved
- Move symbols: _cmd_configure, _emit_team_yaml, _parse_assign_pairs,
  _score_model_for_role, _rank_assignments, _web_enrich, _apply_web_enrichment,
  _TEAM_CAPABILITY_TABLE, _HARNESS_BASE_SCORES, ROLES (ordered list)
- _discover_harnesses() used in-process (no subprocess) as the new seam
- Keep agentic-configure team as back-compat shim delegating via subprocess
- Migrate 6 team tests from test_agentic_configure to test_agentic_team,
  repointing monkeypatch from _run_discover to _discover_harnesses
- Add shim delegation test in test_agentic_configure
- Update both module manifests

Byte-equivalence: agentic-team configure output identical to pre-move
agentic-configure team output. All 114 tests pass.
